requirements.txt
flask
torch
open-clip-torch
Pillow